+++ crates/miden-testing/src/kernel_tests/tx/test_account_interface.rs
@@ -2,7 +2,9 @@ use alloc::vec::Vec;
 
 use assert_matches::assert_matches;
 use miden_lib::note::{create_p2id_note, create_p2ide_note};
+use miden_lib::testing::mock_account::MockAccountExt;
 use miden_lib::transaction::TransactionKernel;
+use miden_objects::Word;
 use miden_objects::account::{Account, AccountId};
 use miden_objects::asset::FungibleAsset;
 use miden_objects::note::{Note, NoteType};
@@ -12,7 +14,6 @@ use miden_objects::testing::account_id::{
     ACCOUNT_ID_SENDER,
 };
 use miden_objects::testing::note::NoteBuilder;
-use miden_objects::{Felt, FieldElement, Word};
 use miden_tx::auth::UnreachableAuth;
 use miden_tx::{
     FailedNote,
@@ -87,12 +88,8 @@ async fn check_note_consumability_custom_notes_success(
     #[case] notes: Vec<Note>,
 ) -> anyhow::Result<()> {
     let tx_context = {
-        let account = Account::mock(
-            ACCOUNT_ID_REGULAR_PUBLIC_ACCOUNT_UPDATABLE_CODE,
-            Felt::ONE,
-            Auth::IncrNonce,
-            TransactionKernel::testing_assembler(),
-        );
+        let account =
+            Account::mock(ACCOUNT_ID_REGULAR_PUBLIC_ACCOUNT_UPDATABLE_CODE, Auth::IncrNonce);
         TransactionContextBuilder::new(account)
             .extend_input_notes(notes.clone())
             .build()?
@@ -135,14 +132,14 @@ async fn check_note_consumability_failure() -> anyhow::Result<()> {
         ChaCha20Rng::from_seed(ChaCha20Rng::from_seed([0_u8; 32]).random()),
     )
     .code("begin push.1 drop push.0 div end")
-    .build(&TransactionKernel::testing_assembler())?;
+    .build(&TransactionKernel::with_kernel_library())?;
 
     let failing_note_2 = NoteBuilder::new(
         sender,
         ChaCha20Rng::from_seed(ChaCha20Rng::from_seed([0_u8; 32]).random()),
     )
     .code("begin push.2 drop push.0 div end")
-    .build(&TransactionKernel::testing_assembler())?;
+    .build(&TransactionKernel::with_kernel_library())?;
 
     let successful_note_1 = create_p2id_note(
         ACCOUNT_ID_REGULAR_PUBLIC_ACCOUNT_IMMUTABLE_CODE.try_into().unwrap(),