+++ crates/miden-testing/src/kernel_tests/tx/test_asset_vault.rs
@@ -2,12 +2,11 @@ use assert_matches::assert_matches;
 use miden_lib::errors::tx_kernel_errors::{
     ERR_VAULT_FUNGIBLE_ASSET_AMOUNT_LESS_THAN_AMOUNT_TO_WITHDRAW,
     ERR_VAULT_FUNGIBLE_MAX_AMOUNT_EXCEEDED,
-    ERR_VAULT_GET_BALANCE_PROC_CAN_ONLY_BE_CALLED_ON_FUNGIBLE_FAUCET,
+    ERR_VAULT_GET_BALANCE_CAN_ONLY_BE_CALLED_ON_FUNGIBLE_ASSET,
     ERR_VAULT_NON_FUNGIBLE_ASSET_ALREADY_EXISTS,
     ERR_VAULT_NON_FUNGIBLE_ASSET_TO_REMOVE_NOT_FOUND,
 };
 use miden_lib::transaction::{TransactionKernel, memory};
-use miden_lib::utils::word_to_masm_push_string;
 use miden_objects::AssetVaultError;
 use miden_objects::account::AccountId;
 use miden_objects::asset::{Asset, FungibleAsset, NonFungibleAsset, NonFungibleAssetDetails};
@@ -22,33 +21,74 @@ use super::{Felt, ONE, Word, ZERO};
 use crate::kernel_tests::tx::ProcessMemoryExt;
 use crate::{TransactionContextBuilder, assert_execution_error};
 
+/// Tests that account::get_balance returns the correct amount.
 #[test]
-fn test_get_balance() -> anyhow::Result<()> {
+fn get_balance_returns_correct_amount() -> anyhow::Result<()> {
     let tx_context = TransactionContextBuilder::with_existing_mock_account().build()?;
 
     let faucet_id: AccountId = ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET.try_into().unwrap();
     let code = format!(
-        "
+        r#"
         use.$kernel::prologue
         use.miden::account
 
         begin
             exec.prologue::prepare_transaction
+
             push.{suffix}.{prefix}
             exec.account::get_balance
+            # => [balance]
 
             # truncate the stack
             swap drop
         end
-        ",
+            "#,
         prefix = faucet_id.prefix().as_felt(),
         suffix = faucet_id.suffix(),
     );
 
-    let process = tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    )?;
+    let process = tx_context.execute_code(&code)?;
+
+    assert_eq!(
+        process.stack.get(0).as_int(),
+        tx_context.account().vault().get_balance(faucet_id).unwrap()
+    );
+
+    Ok(())
+}
+
+/// Tests that asset_vault::peek_balance returns the correct amount.
+#[test]
+fn peek_balance_returns_correct_amount() -> anyhow::Result<()> {
+    let tx_context = TransactionContextBuilder::with_existing_mock_account().build()?;
+    let faucet_id: AccountId = ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET.try_into().unwrap();
+
+    let code = format!(
+        r#"
+        use.$kernel::prologue
+        use.$kernel::memory
+        use.$kernel::asset_vault
+        use.miden::account
+
+        begin
+            exec.prologue::prepare_transaction
+
+            exec.memory::get_acct_vault_root_ptr
+            push.{suffix}.{prefix}
+            # => [prefix, suffix, account_vault_root_ptr, balance]
+
+            exec.asset_vault::peek_balance
+            # => [peeked_balance]
+
+            # truncate the stack
+            swap drop
+        end
+            "#,
+        prefix = faucet_id.prefix().as_felt(),
+        suffix = faucet_id.suffix(),
+    );
+
+    let process = tx_context.execute_code(&code)?;
 
     assert_eq!(
         process.stack.get(0).as_int(),
@@ -78,15 +118,10 @@ fn test_get_balance_non_fungible_fails() -> anyhow::Result<()> {
         suffix = faucet_id.suffix(),
     );
 
-    let process = tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    );
+    let process =
+        tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries());
 
-    assert_execution_error!(
-        process,
-        ERR_VAULT_GET_BALANCE_PROC_CAN_ONLY_BE_CALLED_ON_FUNGIBLE_FAUCET
-    );
+    assert_execution_error!(process, ERR_VAULT_GET_BALANCE_CAN_ONLY_BE_CALLED_ON_FUNGIBLE_ASSET);
 
     Ok(())
 }
@@ -111,13 +146,11 @@ fn test_has_non_fungible_asset() -> anyhow::Result<()> {
             swap drop
         end
         ",
-        non_fungible_asset_key = word_to_masm_push_string(&non_fungible_asset.into())
+        non_fungible_asset_key = Word::from(non_fungible_asset)
     );
 
-    let process = tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    )?;
+    let process =
+        tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries())?;
 
     assert_eq!(process.stack.get(0), ONE);
 
@@ -141,7 +174,7 @@ fn test_add_fungible_asset_success() -> anyhow::Result<()> {
     let code = format!(
         "
         use.$kernel::prologue
-        use.test::account
+        use.mock::account
 
         begin
             exec.prologue::prepare_transaction
@@ -152,13 +185,11 @@ fn test_add_fungible_asset_success() -> anyhow::Result<()> {
             swapw dropw
         end
         ",
-        FUNGIBLE_ASSET = word_to_masm_push_string(&add_fungible_asset.into())
+        FUNGIBLE_ASSET = Word::from(add_fungible_asset)
     );
 
-    let process = &tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    )?;
+    let process =
+        &tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries())?;
 
     assert_eq!(
         process.stack.get_word(0),
@@ -191,7 +222,7 @@ fn test_add_non_fungible_asset_fail_overflow() -> anyhow::Result<()> {
     let code = format!(
         "
         use.$kernel::prologue
-        use.test::account
+        use.mock::account
 
         begin
             exec.prologue::prepare_transaction
@@ -199,13 +230,11 @@ fn test_add_non_fungible_asset_fail_overflow() -> anyhow::Result<()> {
             call.account::add_asset
         end
         ",
-        FUNGIBLE_ASSET = word_to_masm_push_string(&add_fungible_asset.into())
+        FUNGIBLE_ASSET = Word::from(add_fungible_asset)
     );
 
-    let process = tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    );
+    let process =
+        tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries());
 
     assert_execution_error!(process, ERR_VAULT_FUNGIBLE_MAX_AMOUNT_EXCEEDED);
     assert!(account_vault.add_asset(add_fungible_asset).is_err());
@@ -225,7 +254,7 @@ fn test_add_non_fungible_asset_success() -> anyhow::Result<()> {
     let code = format!(
         "
         use.$kernel::prologue
-        use.test::account
+        use.mock::account
 
         begin
             exec.prologue::prepare_transaction
@@ -236,13 +265,11 @@ fn test_add_non_fungible_asset_success() -> anyhow::Result<()> {
             swapw dropw
         end
         ",
-        FUNGIBLE_ASSET = word_to_masm_push_string(&add_non_fungible_asset.into())
+        FUNGIBLE_ASSET = Word::from(add_non_fungible_asset)
     );
 
-    let process = &tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    )?;
+    let process =
+        &tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries())?;
 
     assert_eq!(
         process.stack.get_word(0),
@@ -270,7 +297,7 @@ fn test_add_non_fungible_asset_fail_duplicate() -> anyhow::Result<()> {
     let code = format!(
         "
         use.$kernel::prologue
-        use.test::account
+        use.mock::account
 
         begin
             exec.prologue::prepare_transaction
@@ -278,13 +305,11 @@ fn test_add_non_fungible_asset_fail_duplicate() -> anyhow::Result<()> {
             call.account::add_asset
         end
         ",
-        NON_FUNGIBLE_ASSET = word_to_masm_push_string(&non_fungible_asset.into())
+        NON_FUNGIBLE_ASSET = Word::from(non_fungible_asset)
     );
 
-    let process = tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    );
+    let process =
+        tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries());
 
     assert_execution_error!(process, ERR_VAULT_NON_FUNGIBLE_ASSET_ALREADY_EXISTS);
     assert!(account_vault.add_asset(non_fungible_asset).is_err());
@@ -310,7 +335,7 @@ fn test_remove_fungible_asset_success_no_balance_remaining() -> anyhow::Result<(
     let code = format!(
         "
         use.$kernel::prologue
-        use.test::account
+        use.mock::account
 
         begin
             exec.prologue::prepare_transaction
@@ -321,13 +346,11 @@ fn test_remove_fungible_asset_success_no_balance_remaining() -> anyhow::Result<(
             swapw dropw
         end
         ",
-        FUNGIBLE_ASSET = word_to_masm_push_string(&remove_fungible_asset.into())
+        FUNGIBLE_ASSET = Word::from(remove_fungible_asset)
     );
 
-    let process = &tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    )?;
+    let process =
+        &tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries())?;
 
     assert_eq!(
         process.stack.get_word(0),
@@ -358,7 +381,7 @@ fn test_remove_fungible_asset_fail_remove_too_much() -> anyhow::Result<()> {
     let code = format!(
         "
         use.$kernel::prologue
-        use.test::account
+        use.mock::account
 
         begin
             exec.prologue::prepare_transaction
@@ -366,13 +389,11 @@ fn test_remove_fungible_asset_fail_remove_too_much() -> anyhow::Result<()> {
             call.account::remove_asset
         end
         ",
-        FUNGIBLE_ASSET = word_to_masm_push_string(&remove_fungible_asset.into())
+        FUNGIBLE_ASSET = Word::from(remove_fungible_asset)
     );
 
-    let process = tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    );
+    let process =
+        tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries());
 
     assert_execution_error!(process, ERR_VAULT_FUNGIBLE_ASSET_AMOUNT_LESS_THAN_AMOUNT_TO_WITHDRAW);
 
@@ -397,7 +418,7 @@ fn test_remove_fungible_asset_success_balance_remaining() -> anyhow::Result<()> 
     let code = format!(
         "
         use.$kernel::prologue
-        use.test::account
+        use.mock::account
 
         begin
             exec.prologue::prepare_transaction
@@ -408,13 +429,11 @@ fn test_remove_fungible_asset_success_balance_remaining() -> anyhow::Result<()> 
             swapw dropw
         end
         ",
-        FUNGIBLE_ASSET = word_to_masm_push_string(&remove_fungible_asset.into())
+        FUNGIBLE_ASSET = Word::from(remove_fungible_asset)
     );
 
-    let process = &tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    )?;
+    let process =
+        &tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries())?;
 
     assert_eq!(
         process.stack.get_word(0),
@@ -449,7 +468,7 @@ fn test_remove_inexisting_non_fungible_asset_fails() -> anyhow::Result<()> {
     let code = format!(
         "
         use.$kernel::prologue
-        use.test::account
+        use.mock::account
 
         begin
             exec.prologue::prepare_transaction
@@ -457,13 +476,11 @@ fn test_remove_inexisting_non_fungible_asset_fails() -> anyhow::Result<()> {
             call.account::remove_asset
         end
         ",
-        FUNGIBLE_ASSET = word_to_masm_push_string(&non_existent_non_fungible_asset.into())
+        FUNGIBLE_ASSET = Word::from(non_existent_non_fungible_asset)
     );
 
-    let process = tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    );
+    let process =
+        tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries());
 
     assert_execution_error!(process, ERR_VAULT_NON_FUNGIBLE_ASSET_TO_REMOVE_NOT_FOUND);
     assert_matches!(
@@ -488,7 +505,7 @@ fn test_remove_non_fungible_asset_success() -> anyhow::Result<()> {
     let code = format!(
         "
         use.$kernel::prologue
-        use.test::account
+        use.mock::account
 
         begin
             exec.prologue::prepare_transaction
@@ -499,13 +516,11 @@ fn test_remove_non_fungible_asset_success() -> anyhow::Result<()> {
             swapw dropw
         end
         ",
-        FUNGIBLE_ASSET = word_to_masm_push_string(&non_fungible_asset.into())
+        FUNGIBLE_ASSET = Word::from(non_fungible_asset)
     );
 
-    let process = &tx_context.execute_code_with_assembler(
-        &code,
-        TransactionKernel::testing_assembler_with_mock_account(),
-    )?;
+    let process =
+        &tx_context.execute_code_with_assembler(&code, TransactionKernel::with_mock_libraries())?;
 
     assert_eq!(
         process.stack.get_word(0),