+++ crates/miden-testing/src/mock_chain/auth.rs
@@ -3,15 +3,11 @@
 use alloc::vec::Vec;
 
 use miden_lib::account::auth::{AuthRpoFalcon512, AuthRpoFalcon512Acl, AuthRpoFalcon512AclConfig};
-use miden_lib::transaction::TransactionKernel;
+use miden_lib::testing::account_component::{ConditionalAuthComponent, IncrNonceAuthComponent};
 use miden_objects::Word;
 use miden_objects::account::{AccountComponent, AuthSecretKey};
 use miden_objects::crypto::dsa::rpo_falcon512::SecretKey;
-use miden_objects::testing::account_component::{
-    ConditionalAuthComponent,
-    IncrNonceAuthComponent,
-    NoopAuthComponent,
-};
+use miden_objects::testing::noop_auth_component::NoopAuthComponent;
 use miden_tx::auth::BasicAuthenticator;
 use rand::SeedableRng;
 use rand_chacha::ChaCha20Rng;
@@ -26,7 +22,7 @@ pub enum Auth {
     /// Creates a [SecretKey] for the account, and creates a [BasicAuthenticator] used to
     /// authenticate the account with [AuthRpoFalcon512Acl]. Authentication will only be
     /// triggered if any of the procedures specified in the list are called during execution.
-    ProcedureAcl {
+    Acl {
         auth_trigger_procedures: Vec<Word>,
         allow_unauthorized_output_notes: bool,
         allow_unauthorized_input_notes: bool,
@@ -65,7 +61,7 @@ impl Auth {
 
                 (component, Some(authenticator))
             },
-            Auth::ProcedureAcl {
+            Auth::Acl {
                 auth_trigger_procedures,
                 allow_unauthorized_output_notes,
                 allow_unauthorized_input_notes,
@@ -90,22 +86,9 @@ impl Auth {
 
                 (component, Some(authenticator))
             },
-            Auth::IncrNonce => {
-                let assembler = TransactionKernel::assembler();
-                let component = IncrNonceAuthComponent::new(assembler).unwrap();
-                (component.into(), None)
-            },
-
-            Auth::Noop => {
-                let assembler = TransactionKernel::assembler();
-                let component = NoopAuthComponent::new(assembler).unwrap();
-                (component.into(), None)
-            },
-            Auth::Conditional => {
-                let assembler = TransactionKernel::assembler();
-                let component = ConditionalAuthComponent::new(assembler).unwrap();
-                (component.into(), None)
-            },
+            Auth::IncrNonce => (IncrNonceAuthComponent.into(), None),
+            Auth::Noop => (NoopAuthComponent.into(), None),
+            Auth::Conditional => (ConditionalAuthComponent.into(), None),
         }
     }
 }